- prepare khtml API for 3.0: drop the font size list, introduce a zooming   factor, fix the handling of font zooming. - fixed history state saving / restoring by reverting pmk's commit. This   should fix i.e. Thomas Zander's site - fixed the keyboard event duplication that caused ugly errors like   doubled appearing submit confirmation boxes - fixed pressing "cancel" prevented further submits, as introduced by   George's workaround for the above bug - fixed keyboard event handling again (onkeydown, onkeyup etc) by reverting   pmk's commit and adding the event filter handler (2 liner) - improved accuracy of font size calculation, useful for high-dpi displays - code cleanup
As posted and discussed on kfm-devel, tested by David, George and others.

svn path=/trunk/kdelibs/; revision=143203

+++ khtml/misc/decoder.cpp
@@ -110,8 +110,7 @@ const unsigned char KanjiCode::kanji_map_sjis[] =
 
 /*
  * Maybe we should use QTextCodec::heuristicContentMatch()
- * But it fails detection.
- * It's not useful.
+ * But it fails detection. It's not useful.
  */
 
 enum KanjiCode::Type KanjiCode::judge(const char *str)
@@ -124,7 +123,7 @@ enum KanjiCode::Type KanjiCode::judge(const char *str)
     int euc = 0;
 
     const unsigned char *ptr = (const unsigned char *) str;
-    size_t size = strlen(str);
+    int size = strlen(str);
 
     code = ASCII;
